Move extras folder
